# docs/swagger.yaml
basePath: /v1
definitions:
  dto.ErrorResponse:
    properties:
      error:
        example: message
        type: string
    type: object
  dto.GetSubscriptionHandlerResponse:
    properties:
      end_date:
        type: string
      price:
        type: string
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  dto.ListSubscriptionsHandlerResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      subscriptions:
        items:
          $ref: '#/definitions/dto.SubscriptionItem'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.Period:
    properties:
      end_date:
        type: string
      start_date:
        type: string
    type: object
  dto.StoreSubscriptionHandlerRequest:
    properties:
      end_date:
        type: string
      price:
        type: string
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  dto.StoreSubscriptionHandlerResponse:
    properties:
      id:
        type: string
    type: object
  dto.SubscriptionItem:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: string
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  dto.TotalCostFilters:
    properties:
      service_name:
        type: string
      user_id:
        type: string
    type: object
  dto.TotalCostHandlerResponse:
    properties:
      filters:
        $ref: '#/definitions/dto.TotalCostFilters'
      period:
        $ref: '#/definitions/dto.Period'
      total_cost:
        type: integer
    type: object
  dto.UpdateSubscriptionHandlerRequest:
    properties:
      end_date:
        type: string
      price:
        type: string
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: '-'
  title: Subscription Aggregator
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Get list of subscriptions with filtering and pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: User ID filter (UUID)
        in: query
        name: user_id
        type: string
      - description: Service name filter
        in: query
        name: service_name
        type: string
      - default: start_date
        description: Sort field
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListSubscriptionsHandlerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription
      parameters:
      - description: Subscription data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.StoreSubscriptionHandlerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.StoreSubscriptionHandlerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Delete subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      description: Get subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetSubscriptionHandlerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get subscription
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subscription data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscriptionHandlerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetSubscriptionHandlerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/total-cost:
    get:
      description: Calculate total cost of subscriptions for a specific period with
        optional filters
      parameters:
      - description: User ID filter (UUID)
        in: query
        name: user_id
        type: string
      - description: Service name filter
        in: query
        name: service_name
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TotalCostHandlerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get total cost
      tags:
      - subscriptions
swagger: "2.0"